requirements.txt
argon2-cffi==20.1.0
cffi==1.14.3
click==7.1.2
cssmin==0.2.0
Flask==1.1.2
Flask-Assets==2.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
pycparser==2.20
pyparsing==2.4.7
python-dotenv==0.14.0
six==1.15.0
SQLAlchemy==1.3.19
webassets==2.0
Werkzeug==1.0.1
WTForms==2.3.3
